Updated auth logic to not work with models but get models by email when it's necessary

+++ app/Services/AuthService.php
@@ -3,50 +3,54 @@
 namespace App\Services;
 
 use App\Interfaces\AuthServiceInterface;
+use App\Interfaces\UserServiceInterface;
 use App\Models\User;
+use Exception;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Log;
 use Throwable;
 
 class AuthService implements AuthServiceInterface
 {
+    public function __construct(private UserServiceInterface $userService)
+    {
+    }
+
     public function login(array $credentials): bool
     {
-        if (!Auth::attempt([
-            'email' => $credentials['email'],
-            'password' => $credentials['password']
-        ], $credentials['remember'])) {
+        if (!Auth::attempt($credentials, $credentials['remember'] ?? 0)) {
             return false;
         }
         return true;
     }
 
-    public function logout(): bool
+    public function refreshToken(string $email): string
     {
         try {
-            $this->currentUser()->tokens()->delete();
-            Auth::logout();
-            return true;
-        } catch (Throwable $e) {
-            Log::error($e);
-            return false;
-        }
-    }
-
-    public function currentUser(): ?User
-    {
-        return Auth::user();
-    }
-
-    public function refreshToken(User $user): string
-    {
-        // TODO: userservice ?
-        try {
+            $user = $this->userService->where('email', $email)->first();
+            if (!$user instanceof User) {
+                throw new Exception('Пользователь с указанным email не найден');
+            }
             $user->tokens()->delete();
             return $user->createToken('api-token', ['*'], now()->addHour())->plainTextToken;
         } catch (Throwable $e) {
             Log::error($e->getMessage());
             return '';
+        }
+    }
+
+    public function logout(string $email): bool
+    {
+        try {
+            $user = $this->userService->where('email', $email)->first();
+            if (!$user instanceof User) {
+                throw new Exception('Пользователь с указанным email не найден');
+            }
+            $user->tokens()->delete();
+            return true;
+        } catch (Throwable $e) {
+            Log::error($e);
+            return false;
         }
     }
 }
